requirements.txt
africastalking==1.2.0
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
contextlib2==0.6.0.post1
Django==3.0.7
django-material-admin==1.7.3
django-report-builder==6.3.2
django-reports-admin==1.0.4
django-smart-selects==1.5.6
django-suit==0.3a3
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
mysql-connector==2.2.9
mysqlclient==1.4.6
openpyxl==3.0.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
schema==0.7.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
